requirements.txt
python-dotenv
openai
Pillow
tenacity
scipy
sounddevice
